backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
cloudinary==1.36.0
stripe==5.5.0
PyJWT==2.8.0
Werkzeug==2.3.7